requirements.txt
spoon-ai>=0.5.0
httpx>=0.27.0
python-dotenv>=1.0.0
